+++ Demo/Application/Features/ShopperInsightsViewController.swift
@@ -67,9 +67,6 @@ class ShopperInsightsViewController: PaymentButtonBaseViewController {
         stackView.distribution = .fillEqually
         stackView.translatesAutoresizingMaskIntoConstraints = false
         
-        shopperInsightsClient.sendPayPalPresentedEvent()
-        shopperInsightsClient.sendVenmoPresentedEvent()
-        
         return stackView
     }
     
@@ -85,7 +82,15 @@ class ShopperInsightsViewController: PaymentButtonBaseViewController {
         )
         Task {
             do {
-                let result = try await shopperInsightsClient.getRecommendedPaymentMethods(request: request)
+                let sampleExperiment =
+                    """
+                    [
+                        { "experimentName" : "payment ready conversion" },
+                        { "experimentID" : "a1b2c3" },
+                        { "treatmentName" : "control group 1" }
+                    ]
+                    """
+                let result = try await shopperInsightsClient.getRecommendedPaymentMethods(request: request, experiment: sampleExperiment)
                 // swiftlint:disable:next line_length
                 progressBlock("PayPal Recommended: \(result.isPayPalRecommended)\nVenmo Recommended: \(result.isVenmoRecommended)\nEligible in PayPal Network: \(result.isEligibleInPayPalNetwork)")
                 payPalVaultButton.isEnabled = result.isPayPalRecommended
@@ -97,6 +102,16 @@ class ShopperInsightsViewController: PaymentButtonBaseViewController {
     }
     
     @objc func payPalVaultButtonTapped(_ button: UIButton) {
+        let sampleExperiment =
+            """
+            [
+                { "experimentName" : "payment ready conversion experiment" },
+                { "experimentID" : "a1b2c3" },
+                { "treatmentName" : "treatment group 1" }
+            ]
+            """
+        let paymentMethods = ["Apple Pay", "Card", "PayPal"]
+        shopperInsightsClient.sendPayPalPresentedEvent(paymentMethodsDisplayed: paymentMethods, experiment: sampleExperiment)
         progressBlock("Tapped PayPal Vault")
         shopperInsightsClient.sendPayPalSelectedEvent()
         
@@ -113,6 +128,7 @@ class ShopperInsightsViewController: PaymentButtonBaseViewController {
     }
     
     @objc func venmoButtonTapped(_ button: UIButton) {
+        shopperInsightsClient.sendVenmoPresentedEvent()
         progressBlock("Tapped Venmo")
         shopperInsightsClient.sendVenmoSelectedEvent()